refactor: 그림판 커스텀훅 분리, 로직 일부 수정 #20
